+++ src/i_sound.c
@@ -36,6 +36,8 @@
 ========================================================================
 */
 
+#include <memory.h>
+
 #include "SDL_mixer.h"
 
 #include "c_console.h"
@@ -237,23 +239,36 @@ static void ReleaseSoundOnChannel(int channel)
 }
 
 // Generic sound expansion function for any sample rate.
-static dboolean ExpandSoundData(sfxinfo_t *sfxinfo, byte *data, int samplerate, int length)
+static dboolean ExpandSoundData(sfxinfo_t *sfxinfo, byte *data, int samplerate, int bits, int length)
 {
-    unsigned int        expanded_length = (unsigned int)((((uint64_t)length) * mixer_freq) / samplerate);
+    unsigned int        samplecount = length / (bits / 8);
+    unsigned int        expanded_length = (unsigned int)(((uint64_t)samplecount * mixer_freq) / samplerate);
     allocated_sound_t   *snd = AllocateSound(sfxinfo, expanded_length * 4);
     int16_t             *expanded = (int16_t *)(&snd->chunk)->abuf;
-    int                 expand_ratio = (length << 8) / expanded_length;
+    int                 expand_ratio = (samplecount << 8) / expanded_length;
     double              dt = 1.0 / mixer_freq;
     double              alpha = dt / (1.0 / (M_PI * samplerate) + dt);
 
     if (!snd)
         return false;
 
-    for (unsigned int i = 0; i < expanded_length; i++)
+    if (bits == 8)
     {
-        byte    src = data[(i * expand_ratio) >> 8];
+        for (unsigned int i = 0; i < expanded_length; i++)
+        {
+            byte    src = data[(i * expand_ratio) >> 8];
 
-        expanded[i * 2] = expanded[i * 2 + 1] = (src | (src << 8)) - 32768;
+            expanded[i * 2] = expanded[i * 2 + 1] = (src | (src << 8)) - 32768;
+        }
+    }
+    else
+    {
+        for (unsigned int i = 0; i < expanded_length; i++)
+        {
+            byte    src = ((i * expand_ratio) >> 8) * 2;
+
+            expanded[i * 2] = expanded[i * 2 + 1] = (data[src] | (data[src + 1] << 8));
+        }
     }
 
     // Apply low-pass filter
@@ -271,13 +286,46 @@ dboolean CacheSFX(sfxinfo_t *sfxinfo)
     int             lumpnum = sfxinfo->lumpnum;
     byte            *data = W_CacheLumpNum(lumpnum);
     unsigned int    lumplen = W_LumpLength(lumpnum);
+    int             samplerate;
+    unsigned int    bits = 8;
     unsigned int    length;
 
     // Check the header, and ensure this is a valid sound
-    if (lumplen < 8 || data[0] != 0x03 || data[1] != 0x00)
-        return false;
+    if (lumplen > 44 && !memcmp(data, "RIFF", 4) && !memcmp(data + 8, "WAVEfmt ", 8))
+    {
+        // Make sure this is a PCM format file "fmt " chunk size must == 16
+        if ((data[16] | (data[17] << 8) | (data[18] << 16) | (data[19] << 24)) != 16)
+            return false;
 
-    length = ((data[7] << 24) | (data[6] << 16) | (data[5] << 8) | data[4]);
+        // Format must == 1 (PCM)
+        if ((data[20] | (data[21] << 8)) != 1)
+            return false;
+
+        // Number of channels must == 1
+        if ((data[22] | (data[23] << 8)) != 1)
+            return false;
+
+        samplerate = (data[24] | (data[25] << 8) | (data[26] << 16) | (data[27] << 24));
+        length = (data[40] | (data[41] << 8) | (data[42] << 16) | (data[43] << 24));
+
+        if (length > lumplen - 44)
+            length = lumplen - 44;
+
+        bits = (data[34] | (data[35] << 8));
+
+        // Reject non 8 or 16-bit
+        if (bits != 16 && bits != 8)
+            return false;
+
+        data += 44;
+    }
+    else if (lumplen >= 8 && data[0] == 0x03 && data[1] == 0x00)
+    {
+        samplerate = ((data[3] << 8) | data[2]);
+        length = ((data[7] << 24) | (data[6] << 16) | (data[5] << 8) | data[4]);
+    }
+    else
+        return false;
 
     // If the header specifies that the length of the sound is greater than the length of the lump
     // itself, this is an invalid sound lump
@@ -291,7 +339,7 @@ dboolean CacheSFX(sfxinfo_t *sfxinfo)
     // The DMX sound library seems to skip the first 16 and last 16 bytes of the lump - reason unknown.
 
     // Sample rate conversion
-    return ExpandSoundData(sfxinfo, data + 24, ((data[3] << 8) | data[2]), length - 32);
+    return ExpandSoundData(sfxinfo, data + 24, samplerate, bits, length - 32);
 }
 
 void I_UpdateSoundParms(int channel, int vol, int sep)